fix(asyc-csv): Pass batch size to snuba query
Pass the correct batch size to the snuba query to ensure the number of results returned is as
expected. The current query does not have an explicit limit on the result size so it is using the
default 1000 while the logic makes use of SNUBA_MAX_RESULTS which is set to 10000. This
inconsistency led to the export to include results from index 0-999, then 10000-10999 and so on.